ConsulNamer.Base => LookupCache
